% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{clip_observe}
\alias{clip_observe}
\title{Add an observer to a copy button}
\usage{
clip_observe(input, output, ind, code_expr)
}
\arguments{
\item{input}{The shiny input object.}

\item{output}{The shiny output object.}

\item{ind}{The index of the step.}

\item{code_expr}{An expression (generated by \code{\link[rlang:expr]{rlang::expr()}} that, when
evaluated, returns content to be copied (such as the output from
\code{\link[shinymeta:expandChain]{shinymeta::expandChain()}}.}
}
\description{
Generates an event observer that watches for when a copy button is clicked.
Upon clicking, the \code{code_expr} is evaluated and copied to the clipboard
using \code{\link[clipr:write_clip]{clipr::write_clip()}}.
}
