% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{df_modal_button}
\alias{df_modal_button}
\title{Button to show dataframe modal}
\usage{
df_modal_button(ind, text = "View data")
}
\arguments{
\item{ind}{The index of the step.}

\item{text}{The text to display on the button.}
}
\value{
A \code{\link[shiny:actionButton]{shiny::actionButton()}} with id \verb{df_modal_<ind>}.
}
\description{
Generates a \code{\link[shiny:actionButton]{shiny::actionButton()}} that, when clicked, generates a modal
for the resulting dataset for the specified step. Make sure to set up a
corresponding observer using \code{\link[=df_modal_observe]{df_modal_observe()}}.
}
\examples{
df_modal_button(1)
}
\seealso{
\code{\link[=df_modal_observe]{df_modal_observe()}}, which opens the modal this button triggers.

Other step UI:
\code{\link[=accordion_panel_remove_button]{accordion_panel_remove_button()}},
\code{\link[=select_column_input]{select_column_input()}},
\code{\link[=select_dataset_input]{select_dataset_input()}},
\code{\link[=varname_input]{varname_input()}},
\code{\link[=verbatimTextOutput_copy]{verbatimTextOutput_copy()}}
}
\concept{step UI}
