% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{expandChain_shared}
\alias{expandChain_shared}
\title{Expand code objects with a shared context}
\usage{
expandChain_shared(...)
}
\arguments{
\item{...}{All arguments must be unnamed, and must be one of: 1) calls to
meta-reactive objects, 2) comment string (e.g. \code{"# A comment"}), 3)
language object (e.g. \code{quote(print(1 + 1))}), or 4) \code{NULL} (which will be
ignored). Calls to meta-reactive objects can optionally be \code{\link[=invisible]{invisible()}},
see Details.}
}
\value{
The return value of \code{expandChain()} is a code object that's suitable
for printing or passing to \code{\link[shinymeta:displayCodeModal]{shinymeta::displayCodeModal()}},
\code{\link[shinymeta:buildScriptBundle]{shinymeta::buildScriptBundle()}}, or \code{\link[shinymeta:buildScriptBundle]{shinymeta::buildRmdBundle()}}.
}
\description{
A wrapper for \code{\link[shinymeta:expandChain]{shinymeta::expandChain()}} that uses a shared expansion
context specific to shinypal.
}
