% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{get_int_data}
\alias{get_int_data}
\title{Get an intermediate data object}
\usage{
get_int_data(name)
}
\arguments{
\item{name}{The name of the data object to retrieve.}
}
\value{
The stored reactive; call it (e.g. \code{get_int_data(name)()}) to get
the data. Propagates a \code{req()} failure if \code{name} is not registered.
}
\description{
Retrieves a stored intermediate data reactive by name (the counterpart to
\code{\link[=set_int_data]{set_int_data()}}). Call the returned reactive to obtain the data.
}
\examples{
\dontrun{
df <- get_int_data(input[[paste0("dataset_", ind)]])()
}
}
\seealso{
Other intermediate data:
\code{\link[=get_int_dfs]{get_int_dfs()}},
\code{\link[=set_int_data]{set_int_data()}}
}
\concept{intermediate data}
