% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{include_file}
\alias{include_file}
\title{Set an file to include in the download bundle}
\usage{
include_file(path, name)
}
\arguments{
\item{path}{A path to a file to include in the download bundle.}

\item{name}{A name to store the file as. This should correspond to the name
of the file in the code.}
}
\description{
Set an file to include in the download bundle
}
