% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{select_column_input}
\alias{select_column_input}
\title{Select input to choose a column from a specified dataset}
\usage{
select_column_input(ind, label = "Choose a column:", default = NULL, ...)
}
\arguments{
\item{ind}{The index of the step.}

\item{label}{The label for the select input.}

\item{default}{The default value for the select input.}

\item{...}{Additional arguments passed to \code{\link[shiny:varSelectInput]{shiny::varSelectInput()}}.}
}
\value{
A \code{\link[shiny:varSelectInput]{shiny::varSelectInput()}} tag with id \verb{column_<ind>}.
}
\description{
This should be paired with \code{\link[=select_dataset_input]{select_dataset_input()}}.
}
\examples{
\dontrun{
select_column_input(ind)
}
}
\seealso{
\code{\link[=column_select_observe]{column_select_observe()}}, which keeps this selector in sync.

Other step UI:
\code{\link[=accordion_panel_remove_button]{accordion_panel_remove_button()}},
\code{\link[=df_modal_button]{df_modal_button()}},
\code{\link[=select_dataset_input]{select_dataset_input()}},
\code{\link[=varname_input]{varname_input()}},
\code{\link[=verbatimTextOutput_copy]{verbatimTextOutput_copy()}}
}
\concept{step UI}
