% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{set_int_data}
\alias{set_int_data}
\title{Set an intermediate data object}
\usage{
set_int_data(obj, name)
}
\arguments{
\item{obj}{A reactive data object to store.}

\item{name}{A name to store data object as.}
}
\value{
Called for its side effects; invisibly returns \code{NULL}.
}
\description{
Stores a reactive data object in shinypal's intermediate-data registry
under \code{name}, making it available to later steps and to the assembled
script. Usually called for you by \code{\link[=add_shinypal_data_step]{add_shinypal_data_step()}}.
}
\examples{
\dontrun{
set_int_data(occs, step_varname(ind))
}
}
\seealso{
Other intermediate data:
\code{\link[=get_int_data]{get_int_data()}},
\code{\link[=get_int_dfs]{get_int_dfs()}}
}
\concept{intermediate data}
