% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypal-package.R
\docType{package}
\name{shinypal-package}
\alias{shinypal}
\alias{shinypal-package}
\title{shinypal: Tools to Make Modular and Interactive R Workflow Builders}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides the building blocks for modular, interactive shiny applications that turn point-and-click pipelines into reproducible R and R Markdown code. Using shinymeta, shinypal assembles code modules into a draggable workflow, shows the equivalent code as the user builds it, and lets them export the whole pipeline as a runnable script or rendered report.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/willgearty/shinypal}
  \item \url{http://williamgearty.com/shinypal/}
  \item Report bugs at \url{https://github.com/willgearty/shinypal/issues}
}

}
\author{
\strong{Maintainer}: William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})

Authors:
\itemize{
  \item William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})
}

}
\keyword{internal}
