% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{verbatimTextOutput_copy}
\alias{verbatimTextOutput_copy}
\title{Text element with a button to copy the code to the clipboard}
\usage{
verbatimTextOutput_copy(ind)
}
\arguments{
\item{ind}{The index of the step.}
}
\value{
A \code{\link[htmltools:div]{htmltools::div()}} wrapping a \code{\link[shiny:verbatimTextOutput]{shiny::verbatimTextOutput()}} (id
\verb{code_<ind>}) and a copy \code{\link[shiny:actionButton]{shiny::actionButton()}} (id \verb{copy_<ind>}).
}
\description{
Render a reactive output variable as text within an application
page. Uses \code{\link[shiny:verbatimTextOutput]{shiny::verbatimTextOutput()}} which is usually paired with
\code{\link[shiny:renderPrint]{shiny::renderPrint()}} and provides fixed-width text in a \verb{<pre>}. Make
sure to set up a corresponding observer using \code{\link[=clip_observe]{clip_observe()}}.
}
\examples{
verbatimTextOutput_copy(1)
}
\seealso{
\code{\link[=clip_observe]{clip_observe()}}, which copies the displayed code to the clipboard.

Other step UI:
\code{\link[=accordion_panel_remove_button]{accordion_panel_remove_button()}},
\code{\link[=df_modal_button]{df_modal_button()}},
\code{\link[=select_column_input]{select_column_input()}},
\code{\link[=select_dataset_input]{select_dataset_input()}},
\code{\link[=varname_input]{varname_input()}}
}
\concept{step UI}
