% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{workflow_has_errors}
\alias{workflow_has_errors}
\title{Check whether the workflow has incomplete or errored steps}
\usage{
workflow_has_errors()
}
\value{
A length-one logical.
}
\description{
Returns \code{TRUE} if any step in the current workflow cannot be expanded into
the downloadable script due to failing \code{req()}/\code{validate()} checks.
}
