% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{accordion_panel_remove_button}
\alias{accordion_panel_remove_button}
\title{Accordion panel that includes a remove button}
\usage{
accordion_panel_remove_button(ind, ...)
}
\arguments{
\item{ind}{The index of the step.}

\item{...}{Additional arguments passed to \code{\link[bslib:accordion]{bslib::accordion_panel()}}.}
}
\description{
Accordion panel that includes a remove button
}
