% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_step.R
\name{add_shinypal_step}
\alias{add_shinypal_step}
\title{Add a step to the report, workflow, and the code chain}
\usage{
add_shinypal_step(input, ind, fun_workflow, fun_report, code_chain_list, libs)
}
\arguments{
\item{input}{The input object from the shiny app.}

\item{ind}{The index of the step to be added.}

\item{fun_workflow}{A function that generates the UI elements for the
workflow.}

\item{fun_report}{A function that generates the UI elements for the report.}

\item{code_chain_list}{A list of quoted code bits that will be added to
code_chain().}

\item{libs}{A character vector of R packages required for this step.}
}
\description{
A short description...
}
