% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{column_select_observe}
\alias{column_select_observe}
\title{Keep a \code{varSelectInput} of data.frame column names up-to-date}
\usage{
column_select_observe(input, ind, inputId)
}
\arguments{
\item{input}{The shiny input object.}

\item{ind}{The index of the step.}

\item{inputId}{The id of the \code{\link[shiny:varSelectInput]{shiny::varSelectInput()}} object.}
}
\description{
Keep a \code{varSelectInput} of data.frame column names up-to-date
}
