% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{df_modal_button}
\alias{df_modal_button}
\title{Button to show dataframe modal}
\usage{
df_modal_button(ind, text = "View data")
}
\arguments{
\item{ind}{The index of the step.}

\item{text}{The text to display on the button.}
}
\description{
Generates a \code{\link[shiny:actionButton]{shiny::actionButton()}} that, when clicked, generates a modal
for the resulting dataset for the specified step. Make sure to set up a
\code{\link[shiny:observeEvent]{shiny::observeEvent()}} observer for when the \code{\link[shiny:actionButton]{shiny::actionButton()}} is
clicked, triggering \code{\link[=show_df_modal]{show_df_modal()}}.
}
