% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{df_modal_observe}
\alias{df_modal_observe}
\title{Show a modal with a reactable data.frame}
\usage{
df_modal_observe(input, output, ind, df_name)
}
\arguments{
\item{input}{The shiny input object.}

\item{output}{The shiny output object.}

\item{ind}{The index of the step.}

\item{df_name}{The name of the data.frame to be displayed.}
}
\description{
Show a modal with a reactable data.frame
}
