% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{df_select_observe}
\alias{df_select_observe}
\title{Keep a \code{selectInput} of intermediate data.frames up-to-date}
\usage{
df_select_observe(input, ind)
}
\description{
Keep a \code{selectInput} of intermediate data.frames up-to-date
}
