% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{file_observe}
\alias{file_observe}
\title{Observe a file input to be included in the download bundle}
\usage{
file_observe(input, inputId)
}
\arguments{
\item{input}{The shiny input object.}

\item{inputId}{The id of the \code{\link[shiny:fileInput]{shiny::fileInput()}} object.
of the file in the code.}
}
\description{
Observe a file input to be included in the download bundle
}
