% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_step.R
\name{get_colors}
\alias{get_colors}
\title{Generate a background color and font color for a step based on its index}
\usage{
get_colors(ind)
}
\arguments{
\item{ind}{The index of the step.}
}
\value{
A list containing the \code{color} and \code{background} CSS styles for the
step.
}
\description{
Generate a background color and font color for a step based on its index
}
