% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{get_int_data}
\alias{get_int_data}
\title{Get an intermediate data object}
\usage{
get_int_data(name)
}
\arguments{
\item{name}{The name of the data object to retrieve.}
}
\description{
Get an intermediate data object
}
