% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{get_int_dfs}
\alias{get_int_dfs}
\title{Get the names of all intermediate data.frames for a given step}
\usage{
get_int_dfs(ind)
}
\arguments{
\item{ind}{The index of the step.}
}
\description{
Get the names of all intermediate data.frames for a given step
}
