% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{select_column_input}
\alias{select_column_input}
\title{Select input to choose a column from a specified dataset}
\usage{
select_column_input(ind, label = "Choose a column:", default = NULL, ...)
}
\arguments{
\item{ind}{The index of the step.}

\item{label}{The label for the select input.}

\item{default}{The default value for the select input.}

\item{...}{Additional arguments passed to \code{\link[shiny:varSelectInput]{shiny::varSelectInput()}}.}
}
\description{
This should be paired with \code{\link[=select_dataset_input]{select_dataset_input()}}.
}
