% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{select_dataset_input}
\alias{select_dataset_input}
\title{Select input to choose a shinypal intermediate dataset}
\usage{
select_dataset_input(ind, label = "Choose a dataset:")
}
\arguments{
\item{ind}{The index of the step.}

\item{label}{The label for the select input.}
}
\description{
Select input to choose a shinypal intermediate dataset
}
