% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{set_int_data}
\alias{set_int_data}
\title{Set an intermediate data object}
\usage{
set_int_data(obj, name)
}
\arguments{
\item{obj}{A reactive data object to store.}

\item{name}{A name to store data object as.}
}
\description{
Set an intermediate data object
}
