% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{shinypal_setup}
\alias{shinypal_setup}
\title{Setup shinypal}
\usage{
shinypal_setup(
  input,
  output,
  session,
  modules,
  download_filename = "shinypal_script.zip",
  download_template = "./modules/test_report.qmd"
)
}
\arguments{
\item{input}{The shiny input object.}

\item{output}{The shiny output object.}

\item{session}{The shiny session object.}

\item{modules}{A character vector of paths to independent modules.}

\item{download_filename}{The name of the file to download when the user
clicks the download button.}

\item{download_template}{The path to the Quarto markdown template file that
will be used to generate the report.}
}
\description{
A function to set up the shinypal environment. This function should be
called at the top of the server function of your shiny app.
}
\details{
Each module should have a \code{ui-aux.R} file that defines the UI elements for
that module in the "Workflow" accordion and a \code{server.R} file that defines
the server-side logic for the module.
}
