% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-base.R
\name{shinypal_ui}
\alias{shinypal_ui}
\title{UI for ShinyPal}
\usage{
shinypal_ui(modules)
}
\arguments{
\item{modules}{A character vector of paths to independent modules.}
}
\description{
A function to create the UI for the ShinyPal app. This function should be
called in the UI function of your shiny app.
}
\details{
Each module should have a \code{ui-main.R} file that defines the UI elements for
that module in the "Steps" accordion.
}
