% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{verbatimTextOutput_copy}
\alias{verbatimTextOutput_copy}
\title{Text element with a button to copy the code to the clipboard}
\usage{
verbatimTextOutput_copy(ind)
}
\arguments{
\item{ind}{The index of the step.}
}
\description{
Render a reactive output variable as text within an application
page. Uses \code{\link[shiny:textOutput]{shiny::verbatimTextOutput()}} which is usually paired with
\code{\link[shiny:renderPrint]{shiny::renderPrint()}} and provides fixed-width text in a \verb{<pre>}. Make
sure to set up a \code{\link[shiny:observeEvent]{shiny::observeEvent()}} observer for when the
\code{\link[shiny:actionButton]{shiny::actionButton()}} is clicked.
}
